requirements.txt
mouse
pipwin
instapy
keyboard
pyautogui
python-dotenv
SpeechRecognition